join curvature and decomposition+example_decomposition
