+++ tests/test_curvature.py
@@ -2,7 +2,6 @@ import numpy as np
 import unittest
 import trimesh
 import slam.curvature as scurv
-import slam.curvature_decomposition as scurd
 import slam.generate_parametric_surfaces as sgps
 
 # UTILITIES
@@ -256,8 +255,6 @@ class TestCurvatureMethods(unittest.TestCase):
 
             out += [np.round(np.mean(k_mean_absolute_change), 6)]
 
-        print(out)
-
         # Assert the absolute mean error decreases as we increase the number of
         # points
         assert(sorted(out, reverse=True) == out)
@@ -290,7 +287,7 @@ class TestCurvatureMethods(unittest.TestCase):
 
         mesh_a_save = mesh_a.copy()
 
-        shapeIndex, curvedness = scurd.curvedness_shapeIndex(mesh_a)
+        shapeIndex, curvedness = scurv.curvedness_shapeIndex(mesh_a)
 
         # Non modification
 
@@ -315,6 +312,8 @@ class TestCurvatureMethods(unittest.TestCase):
 
         set_of_tests = [
             # Set = [K, correct shapeIndex, correct curvedness]
+            # For example, on a quadric generated by K=[1,1], the shapeIndex is +-1 at the center
+            # and the curvedness is +-2
             [
                 [1, 1], 1, 2
             ],
@@ -365,7 +364,7 @@ class TestCurvatureMethods(unittest.TestCase):
                     k2_analytic[i] = max(a, b)
 
                 # Decomposition of the curvature
-                shapeIndex, curvedness = scurd.decompose_curvature(
+                shapeIndex, curvedness = scurv.decompose_curvature(
                     np.array((k1_analytic, k2_analytic)))
 
                 # Find the index of the vertex which is the closest to the
@@ -386,11 +385,8 @@ class TestCurvatureMethods(unittest.TestCase):
                         min_m = mg
 
                 # Correctness
-                # ShapeIndex is either 1 or -1 on the vertex at the center of
-                # the mesh
                 computed_shapeIndex = shapeIndex[min_i]
                 computed_curvedness = curvedness[min_i]
-                print(computed_curvedness)
 
                 assert(
                     np.isclose(
